# Conflicts:
#	gorda-be/build/classes/java/main/com/ssafy/gorda/controller/UserController.class
#	gorda-be/build/classes/java/main/com/ssafy/gorda/domain/User$UserBuilder.class
#	gorda-be/build/classes/java/main/com/ssafy/gorda/domain/User.class
#	gorda-be/build/classes/java/main/com/ssafy/gorda/dto/controllerdto/request/LoginUserRequestDto.class
#	gorda-be/build/classes/java/main/com/ssafy/gorda/dto/controllerdto/response/LoginUserResponseDto.class
#	gorda-be/build/classes/java/main/com/ssafy/gorda/repository/UserRepository.class
#	gorda-be/build/classes/java/main/com/ssafy/gorda/service/UserService.class
#	gorda-be/build/tmp/bootJar/MANIFEST.MF
#	gorda-be/build/tmp/compileJava/previous-compilation-data.bin

+++ gorda-be/src/main/java/com/ssafy/gorda/domain/User.java
@@ -34,7 +34,7 @@ public class User {
     private String userNickname;
 
     @Column(name = "user_amount")
-    private int userAmount;
+    private double userAmount;
 
     @Column(name = "user_score")
     private int userScore;
@@ -60,7 +60,7 @@ public class User {
     public User (String userAccount,
                  String userNickname,
                  int userScore,
-                 int userAmount,
+                 double userAmount,
                  String userRole
     )
     {
